ACMS-1081: Move functional test in respective modules.
