+++ modules/acquia_cms_search/tests/src/Functional/AcquiaConnectorTest.php
@@ -1,16 +1,17 @@
 <?php
 
-namespace Drupal\Tests\acquia_cms_tour\Functional;
+namespace Drupal\Tests\acquia_cms_search\Functional;
 
 use Drupal\Tests\BrowserTestBase;
 
 /**
- * Tests the Site Studio Core Form.
+ * Tests the Acquia CMS Connector form.
  *
  * @group acquia_cms
  * @group acquia_cms_tour
+ * @group acquia_cms_search
  */
-class SiteStudioCore extends BrowserTestBase {
+class AcquiaConnectorTest extends BrowserTestBase {
 
   /**
    * {@inheritdoc}
@@ -22,7 +23,8 @@ class SiteStudioCore extends BrowserTestBase {
    */
   protected static $modules = [
     'acquia_cms_tour',
-    'cohesion',
+    'acquia_cms_search',
+    'acquia_connector',
   ];
 
   /**
@@ -41,9 +43,9 @@ class SiteStudioCore extends BrowserTestBase {
   // @codingStandardsIgnoreEnd
 
   /**
-   * Tests the Site Studio Core Form.
+   * Tests the Acquia CMS Connector form.
    */
-  public function testSiteStudioCore() {
+  public function testAcquiaConnector() {
     $assert_session = $this->assertSession();
 
     $account = $this->drupalCreateUser(['access acquia cms tour dashboard']);
@@ -52,25 +54,10 @@ class SiteStudioCore extends BrowserTestBase {
     // Visit the tour page.
     $this->drupalGet('/admin/tour/dashboard');
     $assert_session->statusCodeEquals(200);
-    $container = $assert_session->elementExists('css', '.acquia-cms-site-studio-core-form');
+    // Assert that the expected fields show up.
+    $assert_session->fieldExists('Name');
     // Assert that save and advanced buttons are present on form.
     $assert_session->buttonExists('Save');
-    // Assert that the expected fields show up.
-    $assert_session->fieldExists('API key');
-    $assert_session->fieldExists('Agency key');
-    // Save API key.
-    $dummy_api_key = 'test-key-13fdnj32';
-    $container->fillField('edit-api-key', $dummy_api_key);
-    // Save Agency key key.
-    $dummy_agency_key = 'test-agency-13fdnj32';
-    $container->fillField('edit-agency-key', $dummy_agency_key);
-    $container->pressButton('Save');
-    $assert_session->pageTextContains('The configuration options have been saved.');
-    // Test that the config values we expect are set correctly.
-    $cohesion_api_key = $this->config('cohesion.settings')->get('api_key');
-    $this->assertSame($cohesion_api_key, $dummy_api_key);
-    $cohesion_agency_key = $this->config('cohesion.settings')->get('organization_key');
-    $this->assertSame($cohesion_agency_key, $dummy_agency_key);
   }
 
 }
